requirements.txt
langchain
torch
accelerate
transformers
sentence-transformers
textract
# faiss-gpu
faiss-cpu
tqdm
python-dotenv
jinja2
simsimd